Variable changes and additions
* Sprite now has an imageWidth and imageHeight
* ImageRect width and height have been changed from uint to ushort to
make it easier to convert to int without having to explicitly cast
* Pivot x and y has been changed to a floating point value, should be
assigned a value between 0 and 1 normally
* Transforms now caches cosine and sine values for performance reasons.

+++ Geom/ImageRect.cs
@@ -4,11 +4,11 @@
 	{
 		public int x { get; }
 		public int y { get; }
-		public uint width { get; }
-		public uint height { get; }
+		public ushort width { get; }
+		public ushort height { get; }
 		
 		// --------------------------------------------------------------
-		public ImageRect(int x, int y, uint width, uint height)
+		public ImageRect(int x, int y, ushort width, ushort height)
 		{
 			this.x = x;
 			this.y = y;
