+++ Geom/Pivot.cs
@@ -2,10 +2,10 @@
 {
 	public class Pivot
 	{
-		public int x { get; }
-		public int y { get; }
+		public float x { get; }
+		public float y { get; }
 
-		public Pivot(int x, int y)
+		public Pivot(float x, float y)
 		{
 			this.x = x;
 			this.y = y;
